remove unnecessary deducers
